Update node and mapnik install instructions
Trying to install Node v8 and node-mapnik causes node-mapnik to try
building from source instead of binary. The build fails because
the dependencies for node-mapnik are not complete, and the way to
do that is to install mapnik (not node-mapnik) from source before
installing node-mapnik. I was able to get this working but it takes
a long time to build mapnik from source. Instead using the binary
simplifies the tutorial.